Importeer biologische monsterkenmerken

De functie helpt bij het importeren van biologische monsterkenmerken. De functie leest de kolom datum als Date. Verder wordt er een kolom monster_id toegevoegd op basis van meetpunt, datum en methode.

import_biologie_kenmerken(kenmerken_csv = "data/biologie_kenmerken.csv")

Arguments

kenmerken_csv

Een characterstring met het pad naar het te importeren bestand. Het bestand moet in csv-formaat zijn met ; als scheidingsteken en , als scheidingsteken. Default is "data/biologie_kenmerken.csv". Het is ook mogelijk om een zip-bestand in te lezen waar het csv-bestand in zit.

Value

Een dataframe met biologische kenmerken

Details

Er is enige vrijheid t.a.v. de inhoud van het bestand. De functie verwacht ten minste een kolom datum. Een standaard bestand heeft gewoonlijk de kolommen:

mp Code met de aanduiding van het meetpunt

datum Datum in het format dd-mm-yyyy hh:mm:ss

taxatype Type taxon cf. aquo

kenmerkcode Code van het biologische kenmerk

kenmerknaam Beschrijving van het biologische kenmerk

waarde De numerieke waarde van het kenmerk. Is 0 voor tekstwaarden en klassewaarden

waarde_tekst De tekstwaarde van het kenmerk indien het een tekstkenmerk betreft

waarde_klasse De numerieke klasse-aanduiding

waarde_klassenaam De beschrijving van de klasse

Examples

if (FALSE) {

biologie_kenmerken <- import_biologie_kenmerken(kenmerken_csv = "data/biologie_kenmerken.csv")

}