Importeren van fysisch-chemische data

De functie helpt bij het importeren van fysische chemische data. Rijen zonder waarde worden verwijderd.

import_fys_chem(fys_chem_csv = "data/fys_chem.csv", datumtijd = FALSE)

Arguments

fys_chem_csv

Een characterstring met het pad naar het te importeren bestand. Het bestand moet in csv-formaat zijn met ; als scheidingsteken en , als decimaalteken. Default is "data/fys_chem.csv". Het is ook mogelijk om een zip-bestand in te lezen waar het csv-bestand in zit.

datumtijd

Logical. Of de kolom datum wordt geimporteerd als datum of als datum-tijd. Default is FALSE

Value

Een dataframe met fysisch-chemische meetgegevens.

Details

Er is enige vrijheid t.a.v. de inhoud van het bestand. De functie verwacht ten minste een kolom datum en waarde. Een standaard bestand heeft gewoonlijk de kolommen:

mp Code met de aanduiding van het meetpunt

datum Datum in het format yyyy-mm-dd hh:mm

parnr Unieke nummer van de parameter

par Parametercode, voor de leesbaarheid

eenheid Eenheid, voor de leesbaarheid

detectiegrens Aanduiding als < of >. Bij geen aanduidng krijgt het veld NA

waarde Waarde van de meting

Examples

if (FALSE) {

data <- import_fys_chem()

}