Importeren van meetpunten

De functie helpt bij het importeren van meetpunt-data en bijbehorende meetpuntinformatie.

import_meetpunten(
  meetpunten_csv = "data/meetpunten.csv",
  wide = TRUE,
  code = FALSE,
  tolower = TRUE
)

Arguments

meetpunten_csv

Een characterstring met het pad naar het te importeren bestand. Het bestand moet in csv-formaat zijn met ; als scheidingsteken en , als scheidingsteken. Default is "data/meetpunten.csv". Het is ook mogelijk om een zip-bestand in te lezen waar het csv-bestand in zit.

wide

Logical. Bepaalt of meetpunten in wijd of lang format worden ingelezen. Wide = TRUE is default en resulteert in één regel per meetpunt

code

Logical. Code of omschrijving gebruiken als kolomnamen.

tolower

Logical. Bepaald of alle kolomnamen worden omgezet naar lowercase

Value

Een dataframe met meetpuntinformatie

Details

De functie zet alle kolomkoppen om in lowercase. Er is enige vrijheid t.a.v. de inhoud van het bestand. Een standaard bestand heeft minimaal de kolommen:

mp - Code met de aanduiding van het meetpunt

mpomsch - Omschrijving van het meetpunt

x - X-coordinaat in het RD-stelsel

y - Y-coordinaat in het RD-stelsel

Examples

if (FALSE) { # \dontrun{

meetpunten <- import_meetpunten()

} # }