Importeren van parameters

De functie helpt bij het importeren van parameter-data en bijbehorende parameterinformatie.

import_parameters(parameter_csv = "data/parameters.csv")

Arguments

parameter_csv
Een characterstring met het pad naar het te importeren bestand. Het bestand moet in csv-formaat zijn met ; als scheidingsteken en , als scheidingsteken. Default is "data/parameters.csv". Het is ook mogelijk om een zip-bestand in te lezen waar het csv-bestand in zit.

Value

Een dataframe met parameterinformatie

#' @details Er is enige vrijheid t.a.v. de inhoud van het bestand. Een standaard bestand heeft minimaal de onderstaande kolommen, maar de aquo_kolommen zijn wel wenselijk.

parnr Het parameternummer

par Code voor de parameter, veelal aquo-conform

parnaamlang Uitgebreide naam van de parameter. Geschikt voor weergave

eenheid Eenheid van de parameter

Examples

if (FALSE) {

parameters <- import_parameters()

}