Render R-markdown documenten op de server

De R-server heeft problemen om op het netwerk te renderen. Deze functie omzeilt het probleem door R-markdown documenten eerst lokaal te renderen en daarna to kopieren naar de gewenste locatie.

render_server(rmd_bestand, rmd_map, output_bestand, output_map = rmd_map,
  temp_dir = "C:/JT", ...)

Arguments

rmd_bestand
Naam van het R-markdownbestand

rmd_map
Pad van het R-markdownbestand

output_bestand
Naam van het outputbestand

output_map
Pad naar de outputlocatie

temp_dir
Naam van de tijdelijke map. Default is "C:/JT"

...
Optionele argumenten naar rmarkdown::render

Value

Een gerenderd R-markdownbestand

Note

Deze functie is een "hack" om het renderen op de server zonder problemen te laten verlopen. Het zou wenselijker zijn als direct de functie rmarkdown::render() kan worden aangeroepen.